+++ src/main/kotlin/com/example/sanrio/domain/address/repository/AddressRepository.kt
@@ -1,6 +1,6 @@
-package com.example.sanrio.domain.user.repository
+package com.example.sanrio.domain.address.repository
 
-import com.example.sanrio.domain.user.model.Address
+import com.example.sanrio.domain.address.model.Address
 import com.example.sanrio.domain.user.model.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
